[mojo-tooling] Properly teach `mblack` about ownership transfer operator (`^`) (#35804)
The `mblack` grammar was aware that `^` could be a postscript, but it
would not format it correctly, and in some cases it would crash. This
patch improves the existing hack for the implementation by modifying the
actual black logic to correct emit spaces around postscript `^`
operators, and also teaches the grammar about the case when `^` is
followed by `.` (i.e. attribute access).

All changes outside the `mblack` implementation are NFC reformatting.

Closes [Internal Link]

modular-orig-commit: 60025badf6032b1a0a561f83f1b8e55870a7a41b