[Docs] Add tutorial to the development guide (#35830)
[Docs] Add tutorial to the development guide

This adds a end-to-end tutorial to the development guide for users that
struggle to comprehend how everything fits together. Also some
modifcations to the rest of the doc, to fill in any gaps in knowledge.

MODULAR_ORIG_COMMIT_REV_ID: 6a58f91ef5ba91653dfe44cdb2c550c142c5a319