[mojo][repl] Fix erroneous destructor call emission in generated REPL source (#35595)
The REPL is doing dirty tricks with variables, turning them into heap
allocations and tracking them in indirect structs. However,
CheckLifetimes
is onto its tricks, which causes it to reject these as incorrect values.

The right solution is to make more invasive changes to the REPL, but
that isn't in the short term plans.  Workaround this by introducing a
horrible hack (tm) op `RefFromPointerUntrackedOp` that creates a
reference that isn't tracked by CheckLifetimes.

This pokes a hole in our nice reference system which makes me very
sad.  A comment saying "don't use this" will prevent other people from
using this for other things ... right???

---------

Co-authored-by: Chris Lattner <clattner@nondot.org>

modular-orig-commit: 7a6b7339522d2657f72d7e8a3a5b51e4d299167f