[mojo-lang] Expand `VariadicPack` to take parametric trait base (#36319)
This patch is a big step towards enabling VariadicPack to model
packs with a non-AnyType bound (e.g. all members must be Stringable).
It expands `VariadicPack` with a new `element_trait` member that
indicates what all the elements are, and teaches the parser to
pass down the element type.

The magic enabling this is support for metatype bound type
expressions which is almost working, but not quite there yet. As
such, this is another step, but isn't enough to declare success,
which is why the integration test isn't using it yet. :-/

MODULAR_ORIG_COMMIT_REV_ID: d45b597eef0f58fc20770c2b7ec860216e9ecc6c