[stdlib] Check compiler versions before building
There's often breaking changes in the compiler that mandate an update to the
latest `nightly/mojo` compiler in order to work with the latest standard
library.  Currently, if a user installs a newer `nightly/mojo` compiler but
hasn't updated the library source code, they'll get pages of bad errors
depending on the breaking change, often "cryptic" MLIR errors.  This is not a
great user experience, and does not tell contributors what they can/should do.

To improve this user experience, check the current installed compiler version
in the script when building the standard library.  If there is a mismatch
between the expected compiler version (stored in a file in the repo) and the
installed compiler, give a good error message telling the user to update their
`nightly/mojo` compiler.

Note:
- An alternative approach is to query to find the "latest compiler version
  released" and check against that, but that has the caveat that just because
  internally we publish a new `nightly/mojo` package, if someone hasn't updated
  their local copy of the standard library source code, they shouldn't be
  prevented from building just because something out-of-band changed/got
  updated.  This approach keeps the local flow working until they update their
  copy of the library source code.
- We'll probably want to rewrite these bash scripts into Python for a bit better
  maintenance soon.