[mojo-lsp] Fix hover snippet for functions with functional args/params/result (#36391)
The LSP was using `->` as a heuristic for the separation between the
arguments and results of a function signature, which isn't correct in
the presence of functional types. This PR fixes the rendering to
properly track the offset of a return, to ensure the hover is always right.

Closes https://github.com/modularml/mojo/issues/1215
Closes https://github.com/modularml/mojo/issues/1949

MODULAR_ORIG_COMMIT_REV_ID: ed82e2e607cda9c8d7a018d0d545a453443962c3