[stdlib] Add `Dict.update()` (#36677)
[Here](https://docs.python.org/3/library/stdtypes.html#dict.update) is the
python documentation about this method.

There should be another overload: `fn update(self, **kwargs: V)`
which we can't do yet since it would only work with keys that are
strings and we don't yet have conditional conformance (see
[Internal Link]

This is why we force the argument to be positional-only. If we allowed it
to be passed by keyword, it would break when we will introduce
the other overload of `Dict.update`.

Upstream commit:
[Internal Link]

modular-orig-commit: d3fcf8609c6d97b4ea4d2bf866987d89e9e30a86