[stdlib] Add `is` and `isnot` functions to `Optional` (#36674)
Just as the title says, it's a very common idiom in python, is even
recommended by PEP8:
https://peps.python.org/pep-0008/#programming-recommendations.

Comparisons to singletons like `None` should always be done with is or
is not, never the equality operators.
Also, beware of writing if x when you really mean if x is not None –
e.g. when testing whether a variable or argument that defaults to None
was set to some other value. The other value might have a type (such as
a container) that could be false in a boolean context!

While the warning does not apply to Mojo, Python users are quite used to
it and it's a small quality of life improvement which is inexpensive to
support.

Upstream commit: https://github.com/modularml/mojo/commit/6308519882f67b9dc7e7ea2ad584e3f35da2e009

MODULAR_ORIG_COMMIT_REV_ID: 0961d7b5804b574df61806b8c433471e72cfcd67