[mojo-lang] Enable `VariadicPack` for borrowed memory packs (#36680)
This is the big one: `VariadicPack` is now ready to take on the
heavy load for all non-register-passable packs, which notably
includes `print` and `String.join`.  This means deleting our
old friend `_StringableTuple`, replacing the logic with stuff
that is much simpler.

These examples use capturing values, so switch the
`VariadicPack.each` method to work with a capturing closure

This patch fixes a ton of bugs, because variadic packs of
memory-only type now work!

modular-orig-commit: 436b86a7496642a6a8f4c4921ef7234436618123