[Scripts] Add override for compiler version checking
Add an override ability for doing the compiler version checking.  This is opt-in
based on the environment variable `MOJO_OVERRIDE_COMPILER_VERSION_CHECK`.  By
default, we still do the compiler version checking.  Internally, this allows us
to test to make sure the workflow upstream works for unreleased compilers that,
by definition, would have a different compiler version than the one in the
`COMPATIBLE_COMPILER_VERSION` file.

Signed-off-by: Joe Loser <joe@modular.com>