[mojo-stdlib] Move `llvm_intrinsic` uses off of `Tuple`. (#37009)
Multiresult uses of `llvm_intrinsic` are currently using Tuple to
represent the result list, which conveniently works out given how
lowering happens.  However, `Tuple` will move to a new
representation and needs to be memory only to do it.  Introduce a
stopgap solution to get these uses off of it.

modular-orig-commit: 9ff10499f5787c411b2b0b8471acc424aa03eb76