[mojo-stdlib] Remove the need type in `yourIntTuple.get[1, Int]()` (#37014)
It is very common to work with tuples of known element types along
with statically known indexes.  The Mojo type system is strong
enough to propagate this through parametric types.  Add an overload
of get that doesn't take a type parameter - but keep that form for
when a rebind is needed.

MODULAR_ORIG_COMMIT_REV_ID: 60020ed2c3837dd3161473638083beb90d49adf8

Signed-off-by: Sriram Vadlamani <sriram.vadlamani@proton.me>