[mojo-stdlib] Remove `Movable` requirement from `AnyPointer` (#37059)
This removes the `Movable` requirement from `AnyPointer`, allowing
it to be adopted in place of `Pointer`.  This requires moving a
few methods to being global functions.  While doing this, rename
them for clarity.  This also removes the `Movable` requirement from
`abort()`.

This adds a `destroy_pointee` function which is more clear, but also
more efficient than the `_ = ptr.takeValue()` idiom: it just calls del,
not moveinit+del.

modular-orig-commit: 2ddef633fe90115c713b48b4be556972ec5c20db