[mojo-stdlib] Rewrite `Tuple` to support memory-only elements. (#37032)
With enough blocking work out of the way, we can finally rewrite
`Tuple` to support memory only elements.  It does so by having a
`!kgen.pack` storage container, and manually copying/moving the
elements provided to its copy/move constructor into its storage,
and explicitly destroys them when it is done.

This is a great step forward, but hasn't been fully tested yet,
that will come in a follow-on.

modular-orig-commit: 88306e320be862b8caffebe3749a83f591ec3332