[mojo-stdlib] Sundry improvements to pointer/reference. (#37138)
This makes use of `AnyPointer`'s new-found generality to `AnyType`,
allowing use of it to simplify a bunch of code.

As part of this, we remove `emplace_ref_unsafe` and the
`destroy_element_unsafe` method in favor of the `AnyPointer`
versions.

MODULAR_ORIG_COMMIT_REV_ID: 014e0bced52a154aceac00eb64d11f8f501822fb

Signed-off-by: Sriram Vadlamani <sriram.vadlamani@proton.me>