[stdlib] Move `parallel_memcpy` to `algorithm` package (#37432)
Currently, the `buffer` package depends on the `algorithm` package which depends
on the `buffer` package, which creates a package cycle.  Package cycles are bad
for a lot of reasons.  So, to fix the cycle, just move `parallel_memcpy`
function into the `algorithm` package.  Previously, the `buffer` package only
depended on the `algorithm` package for one function: `sync_parallelize` which
was only used in the `parallel_memcpy` function.  So, this is the easiest
solution.

Fixes [Internal link]

MODULAR_ORIG_COMMIT_REV_ID: 0bb8e52a2ac532133149f0a954692309777009d4