[tools] Add a `pre-commit` for `mojo format` and license check (#2193)
We currently rely on contributors applying themselves the `mojo format -l 80` and thinking themselves about the licence check. This PR adds a pre-commit, which, once installed, will do the mojo format and license check automatically before each commit.

This is optional for contributors but useful, since, if later we add more steps in the pre-commit, they will automatically run (both on the contributors' laptops and in the CI).

This should save some time for the maintainers as less mistakes will be made and the CI will check more stuff. Here, the CI will fail if the license is not present on every file.

I also simplified the contributing guide, notably removed the section about whitespaces, and line length which is already enforced by mojo format.

This also means that I had to fix all files missing the license (3) and all Mojo files which had an incorrect format (2)

Signed-off-by: Gabriel de Marmiesse <gabrieldemarmiesse@gmail.com>