[mojo][docs] Generate separate page for struct/trait/function (#37536)
Creates a separate template for module/package summaries,
structs/traits, and functions. This provides a much more scanable
overview on each module page, so it's easy to see all stucts, traits,
and functions with their summaries.

This should help disambiguate search results because a module with
several structs currently gets just one search result and it's usually
not clear that is the page you want. Instead having a separate URL with
a distinct title makes it much easier to search and find the API you're
looking for. It also makes it a lot easier to identify/distinguish
traits from structs.

See stage here:
[Internal Link]

resolves: [Internal Link]

modular-orig-commit: 730346773d4ce2a405a3363709b5d6a117cc481a