[docs] Update links to stdlib as per new file structure (#37595)
This follows the mojodoc structure changes from
[Internal Link]

Example...
Previously:
https://docs.modular.com/mojo/stdlib/builtin/int.html#intableraising
Now:
https://docs.staging.modular.com/mojo/stdlib/builtin/int/intableraising

There are bound to be some mistakes in this find/replace but this mostly
looks effective. But there's also been a lot of refactoring going on in
the stdlib, so I already noticed during my review that some of these
links don't work due to that refactoring and not these URL changes. I'll
run a link checker post-publish to catch any issues.

modular-orig-commit: efce16f59f2fff6f8437774f0474dd17335fd2f9