[mojo-stdlib] Remove address space support from `Dict` (#37651)
Address spaces are really important for low-level types like the
pointers and things like NDBuffer built on top of them, but aren't
something that Mojo supports in full generality.  In particular,
argument conventions like 'borrowed' and 'inout' are implicitly
references with address space zero.

The consequence of this is that you can't delete one of these
(because the self of `__del__` is always in address space zero)
which means that address spaces are irrelevant to things using
traits.  It is theoretically possible we'll re-evaluate this in
the future, but until then, lets not address-spacify everything:
it just adds complexity.

modular-orig-commit: 7ba0d6225a00cd36552092c8ff05ebb595d1d7ea