[mojo-lang] Fix non-trivial register variadics. (#37671)
VariadicList is implemented with AnyRegType so it can't invoke
copy constructors.  This means we can't use it for non-trivial types
like PythonObject.  Switch variadics of non-trivial register type to
use VariadicListMem instead: we'd eventually like to remove
VariadicList but are not quite ready for that.

This fixes variadics with things like PythonObject and
Fixes #37362 and Fixes [Internal Link]

modular-orig-commit: fd3cdc841773c009ccce94110e722961fa2a1b82