[mojo-stdlib] Add file read method that reads data into `DTypePointer` (#37441)
Allows the user to pass in an allocated `DTypePointer` to read data into the
memory location from the position set by the seek pointer. Previously could only
receive back a `String` or `List[Int8]`. The use-case this supports is `llm.c` which makes
heavy use of C's `fread` and offsetting pointers while reading chunks of data into a
base memory location. The API here is similar to `fread`, but fits in with Python having
a `read` method on a `FileHandle`, it also matches the rest of Mojo where `size` is the
number of elements, and you don't have to pass in the number of bytes for each element
like in C.

modular-orig-commit: cc52fb7e604c72c803e074b30a18a1c8b6de626a