[External][stdlib] Fix assert_equal[SIMD]() not raising if any elements are equal (#2279) (#38063)
To reproduce the bug:
```mojo
with assert_raises():
    assert_equal(SIMD[DType.uint8, 2](1, 1), SIMD[DType.uint8, 2](1, 2))
```

This is because using `!=` returns `[False, True]` then, when using
`.__bool__()` on this, since `__bool__()` is equivalent to a
`reduce_and()`, it returns False, and the error is not triggered.

Long story short, doing `if a != b` doesn't do what one would expect.
This bug is not present with `assert_not_equal()`.

Signed-off-by: gabrieldemarmiesse <gabrieldemarmiesse@gmail.com>

mojo-orig-commit: ee5dfe791effa9def469001e9cd5933996b96c6f
MODULAR_ORIG_COMMIT_REV_ID: 55d43d1473a9005ab1157851c60251a4f16a036a

Signed-off-by: Sriram Vadlamani <sriram.vadlamani@proton.me>