[External][stdlib] Remove unnecessary declaration in `base64` (#2275) (#38179)
`i` was an unnecessary and unclear declaration that could be substituted
for `end`. This is not a behavioral change.

Reasoning:
- `i` was set to be equivalent to `end`, but was never used in a way in
which it could not simply be substituted for `end`.
- `i` is not descriptive, and `end` does a better job describing it's
purpose. `i` didn't have a comment either, also making it less
descriptive than `end`.
- `var i` was used almost immediately after a different `i` was used as
a counter in a for loop, and it could be confusing to use two variables
with different purposes and the same name immediately after one another.
- Removing this declaration could result in a small performance
improvement because an assignment is removed, but at the very least the
code is simpler without this declaration.

Signed-off-by: Liam Swayne <liamtswayne@gmail.com>

mojo-orig-commit: 0a884b7941303e41d77e889101c34783c9e3a975

Co-authored-by: Liam Swayne <liamtswayne@gmail.com>
MODULAR_ORIG_COMMIT_REV_ID: 8f33db92dc2ff543894fc2edb4cc49bc7759b68c