[mojo-lang] Continue refactoring getitem/setitem/refitem emission. (#38248)
This continues to move around the logic for this to make it so
index expressions are only emitted after at least one candiate
set is found that will use it.  This isn't used yet but will in
a future patch.

This also makes the logic hard default to refitem implementations
if present, so adjust some weird hybrid code in LegacyPointer to
be more modern.

modular-orig-commit: 61f5e0ec9d0fb2be4f2a70e22ae4f78949cbf2f3