[mojo-stdlib] Generialize `debug_assert` to take `Stringable` (#38540)
There is nothing in the implementation that requires a `StringLiteral`;
in fact, we actually need to instantiate a string from it anyway to
concatenate for warnings. This patch also simplifies the implementation
and renames some test files to conform to naming conventions. A
redundant overload is also removed.

MODULAR_ORIG_COMMIT_REV_ID: 252c211943a50ad2c84bcb169626ca2a4b0104d7

Signed-off-by: Sriram Vadlamani <sriram.vadlamani@proton.me>