[External] [stdlib] Allow `String()` to be created from `UInt8` ptr and list too (#38859)
[External] [stdlib] Allow `String()` to be created from `UInt8` ptr and
list too

This is related to https://github.com/modularml/mojo/issues/2317.

It allows a graceful and slow migration to UInt8 as the primary byte
type. This allows everyone to give whatever (Int8 or UInt8) to the
String struct and it just works™

When everything uses Uint8, we can remove the old methods that work with
Int8, and the migration will be complete.

Co-authored-by: gabrieldemarmiesse <gabrieldemarmiesse@gmail.com>
Closes modularml/mojo#2448
MODULAR_ORIG_COMMIT_REV_ID: d12649d6dfa178eb756d56df563ad155b5240093

Signed-off-by: Lukas Hermann <lukashermann28@gmail.com>