[External] [stdlib] Complete most of the remaining methods of `Set` (#38990)
[External] [stdlib] Complete most of the remaining methods of `Set`

I've been working on this for a while and I believe this PR is a big
step toward reaching the point that our collection datatypes are almost
identical to Python. Long story short, First, let me introduce you to
the new methods and operators that are being implemented with this PR:

## Methods
1- `issubset`
2- `isdisjoint`
3- `issuperset`
4- `symmetric_difference`
5- `symmetric_difference_update`
6- `discard`
7- `clear`

## Operators
1- `__le__`
2- `__lt__`
3- `__ge__`
4- `__gt__`
5- `__xor__`
6- `__ixor__`

## Renamed methods
- `remove_all` renamed to `difference_update`

## The EDGE CASES:
The methods are designed to replicate the same functionality we get from
them in Python. So, This PR assures that we have no methods/operators
left that are not coded yet. Cool, right? But there are still some
points that must be considered. I list them below:

1- The only method not implemented yet is `copy`. I haven't implemented
this because of the differences between Python and Mojo and the related
concepts of shallow/deep copy. Any suggestion is welcomed.

2- I have observed `pop` method has slightly different functionality
than the Pythonic version. In Python, `pop` method works by randomly
deleting an item, which is not what we have in Mojo's sets. We can keep
the current functionality and also add the Pythonic one. They can be
declared all in one single method but a default argument can determine
if the user wants the random poping or the other.

3- In Python, most of the methods that accept another set (e.g.
`issubset`), work with variadic arguments which means you can pass more
than one set/collection to them. The newly added and the previously
implemented methods do not still support such a thing. I'm deciding
whether to add them here.

Co-authored-by: ARVIN DAVOUDI <arvindavoudi242@gmail.com>
Closes modularml/mojo#2456
MODULAR_ORIG_COMMIT_REV_ID: 2f00350ea0b211b5598207703f6168f8585f83f9