[External] [stdlib] Remove unnecessary arguments and overloads in print() (#39173)
[External] [stdlib] Remove unnecessary args and overloads in print()

The print function was using overloads and `rest.each` to be able to
generate all the `_put` call necessary at compile time. It's possible to
do the same thing without overloads and also with less arguments. This
will increase maintainability and make the signature easily
understandable in the docs.

As discussed with Billy and Walter offline, disable the broken debug-info tests
that result due to the code-gen differences of `print` and how we manage
lifetime tracking in the debug-info pipeline.

Co-authored-by: gabrieldemarmiesse <gabrieldemarmiesse@gmail.com>
Closes modularml/mojo#2463
MODULAR_ORIG_COMMIT_REV_ID: 8aff0f6735bdcd0ad29928a7197ea4d26ef24a5c