[mojo-stdlib] Massively simplify parametric mutability reference usage (#39369)
Now that we can declare `self` to be a `Reference` and not just a
`!lit.ref` we can use parameter inference to deduce the mutability
and lifetime and use `_` expressions to avoid having to explicitly
declare them.

This radically simplifies the usage across the stdlib.

MODULAR_ORIG_COMMIT_REV_ID: 123cd20425d4a4adcd609cc74a254e8ab116f29f