Revert "[stdlib] Add move constructor to `Atomic` (#36718)" (#39147)
This reverts commit 7654fad7242cdb9d88759558a1ba7ca360889144. Since
atomic operations are shared across threads and defined by memory
identity, it's not a great idea to define a move constructor for
`Atomic`. Deleting the move constructor is the element of least
surprise.

In the future, we may want to make it easier to get the value out of an
atomic, but it shouldn't be implicitly movable - this was a footgun
before this patch.

MODULAR_ORIG_COMMIT_REV_ID: 160d1d6e8afc52fb5568dea06733730cdff8d9ca