[External] [stdlib] Make `repr()` work with  `DType` (#39511)
[External] [stdlib] Make `repr()` work with  `DType`

This PR is a part of https://github.com/modularml/mojo/pull/2482. This is because https://github.com/modularml/mojo/pull/2482 has become non-trivial
to do due to the introduction of the Formatter struct. I'll need time to
rework the `__repr__` for the SIMD struct. In the meantime we can merge
the implementation for DType.

Co-authored-by: Gabriel de Marmiesse <gabriel.demarmiesse@datadoghq.com>
Closes modularml/mojo#2581
MODULAR_ORIG_COMMIT_REV_ID: 7ac170a65d45211aa0a18a204cb17889802b86c4

Signed-off-by: Lukas Hermann <lukashermann28@gmail.com>