[stdlib] Change `Bool` to be backed by an `i1` (NFC)
This PR changes the `Bool` builtin type to be backed by an `i1` instead
of a `!pop.scalar<bool>`. The main benefit of this is that it makes the
`__mlir_i1__` implementation trivial, which significantly reduces IR
bloat
and the amount of back-and-forth casting to `!pop.scalar<bool>`.

MODULAR_ORIG_COMMIT_REV_ID: 4ab8172a81c6f185059b5573d5518f02db2466fc

Signed-off-by: Lukas Hermann <lukashermann28@gmail.com>