[CI] Add timeouts to workflows
On Ubuntu tests, we're seeing some non-deterministic timeouts
due to a code bug (either in compiler or library) from a recent
`nightly` release.  Instead of relying on the default GitHub timeout
of 6 hours for all of our workflows, add a sensible timeout
for each of our jobs, so they end early rather than spinning
for many hours.

Signed-off-by: Joe Loser <joeloser@fastmail.com>