[stdlib] Fix start/stop semantics in `List.index`
The semantics of these should match the semantics of slice boundaries.
This means that we shouldn't do any explicit bounds checking: if an
equivalent slice yields an empty sublist, we should just fail with the
same error as if the element was not found. This patch also removes the
extra check for empty list in this method, allowing more convenient and
python error propagation.

MODULAR_ORIG_COMMIT_REV_ID: 8a695f69dbcfed8db699dec723effd26a407b435

Signed-off-by: Lukas Hermann <lukashermann28@gmail.com>