[External] [stdlib] Guard `SIMD.cast()` against equal types (#40022)
[External] [stdlib] Guard `SIMD.cast()` against equal types

Checks if types are already equal when calling `SIMD.cast()`.
This avoids unnecessary code when generically casting simd vectors.

But, this could be used in place of rebind when you know the types are
already equal.
if that's not what we want, then maybe we should constrain the types to
be different, instead of doing this.

Co-authored-by: Helehex <Helehex@gmail.com>
Closes modularml/mojo#2669
MODULAR_ORIG_COMMIT_REV_ID: 737b3d0a79e304d131a79af7e21cb985dc4267ad

Signed-off-by: Lukas Hermann <lukashermann28@gmail.com>