[External] [stdlib] Fix wrong slice check in span (#40212)
[External] [stdlib] Fix wrong slice check in span

Fix part of https://github.com/modularml/mojo/issues/2687

both the start and the end of span can be `len(original_string)`, that's
not an issue. In the case where the start is equal to the lenght of th
string, it just gives an empty span. This is similar to the slicing
`my_string[:0]` which also gives an empty span.

Co-authored-by: Gabriel de Marmiesse <gabriel.demarmiesse@datadoghq.com>
Closes modularml/mojo#2726
MODULAR_ORIG_COMMIT_REV_ID: 1e08d3d29352ae43b7e833d36f5a3acfa7ebe765