[External] [docs] Add a note about pull request sizes in the contributing guide (#40021)
[External] [docs] Add a note about pull request sizes in the
contributing guide

We're seeing bigger and bigger PRs. It would be good for the overall DX
if contributors were aware of the effect of huge PRs on the whole PR
process.

Reviewers will be able to reference this section when faced with a pull
request that can be split into multiple ones. I will avoid discussions
between the reviewer and the contributor about who is the laziest of the
two (I've already witness this situation in another project haha)

Co-authored-by: Gabriel de Marmiesse <gabriel.demarmiesse@datadoghq.com>
Closes modularml/mojo#2663
MODULAR_ORIG_COMMIT_REV_ID: 8e4d8bc88530c05b370dcb86c6c5c5eafd215729