Fix docsite build, with a period
This is a bug with our website, but this is also a workaround to fix it.
The widget that builds the stdlib module index page currently requires
this sentence to end in a period.

MODULAR_ORIG_COMMIT_REV_ID: be9c32611dc7f55bdd8fd495b5f1b3e923140c9a