[mojo] Fix some invalid `AnyType -> AnyRegType` bindings (NFC)
Binding a generic Mojo type to `AnyRegType` is invalid, because the
latter assumes that the type is register passable (and trivial -- it is
incorrectly named). The compiler allow this for legacy reasons but we
intend to make it an error so the code is semantically correct and
safety is enforced.

This PR starts transitioning some violations of this:

- `get_sizeof` and `get_alignof` are changed to take an `AnyType`
- some invalid rebinds of generic types were fixed

MODULAR_ORIG_COMMIT_REV_ID: a0b9e17e8786c1ea418bf18091bb4e1b5f062583