[External] [stdlib] Add `String.__iter__()` and fix `String.isspace()` (#41248)
[External] [stdlib] Add `String.__iter__()` and fix `String.isspace()`

Add an iterator for String that takes multi byte utf8 encoding into
account (other encodings remain as `TODO`).

Used a comparison in the utf8 bit range to know if a byte is a 1 byte,
continuation byte, a 2 byte start, a 3 byte start, or a 4 byte start.

The algorithm is made to be faster for languages with mostly ASCII
characters, because the amount of continuation bytes in a string are
counted in the Iterator constructor. That way when all of the characters
with continuation bytes have been iterated over (the continuation_bytes
variable becomes 0), the `_utf8_byte_type` function does not need to be
called and a Reference of length 1 is directly returned on every
subsequent iteration.

The iterator returns a StringSlice of the corresponding byte length
(without null terminator length).

ORIGINAL_AUTHOR=martinvuyk
<110240700+martinvuyk@users.noreply.github.com>
PUBLIC_PR_LINK=modularml/mojo#2868

Co-authored-by: martinvuyk <110240700+martinvuyk@users.noreply.github.com>
Closes modularml/mojo#2868
MODULAR_ORIG_COMMIT_REV_ID: 7ca8fce3da1218c13672c9c242a0fc0e89709a77