[External] [stdlib] Simplify iterator logic in `dict.mojo` (#41265)
[External] [stdlib] Simplify iterator logic in `dict.mojo`

There is no need to increment or decrement the index in both branches of
the if. We just need to do it once, before the if statement, since
`self.index` is not used afterwards.

It also makes it easier to understand the purpose of the if statement.

Co-authored-by: Gabriel de Marmiesse <gabriel.demarmiesse@datadoghq.com>
Closes modularml/mojo#2934
MODULAR_ORIG_COMMIT_REV_ID: 425f8f41b2d9d7971c2971891a718d8cc0e6d7aa